+++ lib/ace/http2/frame/2_priority.ex
@@ -31,4 +31,10 @@ defmodule Ace.HTTP2.Frame.Priority do
     exclusive = if frame.exclusive, do: 1, else: 0
     <<5::24, 2::8, 0::8, 0::1, frame.stream_id::31, exclusive::1, frame.stream_dependency::31, frame.weight::8>>
   end
+
+  defimpl Inspect, for: Ace.HTTP2.Frame.Priority do
+    def inspect(%{stream_id: stream_id, stream_dependency: stream_dependency, weight: weight, exclusive: exclusive}, _opts) do
+      "PRIORITY(stream_id: #{stream_id}, stream_dependency: #{stream_dependency}, weight: #{weight}, exclusive: #{exclusive})"
+    end
+  end
 end