+++ lib/ace/http2/frame/5_push_promise.ex
@@ -36,4 +36,10 @@ defmodule Ace.HTTP2.Frame.PushPromise do
     flags = <<0::4, padded_flag::1, end_headers_flag::1, 0::2>>
     <<length::24, 5::8, flags::binary, 0::1, frame.stream_id::31, 0::1, frame.promised_stream_id::31, frame.header_block_fragment::binary>>
   end
+
+  defimpl Inspect, for: Ace.HTTP2.Frame.PushPromise do
+    def inspect(%{stream_id: stream_id, promised_stream_id: promised_stream_id, end_headers: end_headers}, _opts) do
+      "PUSH_PROMISE(stream_id: #{stream_id}, promised_stream_id: #{promised_stream_id}, end_headers: #{end_headers})"
+    end
+  end
 end